requirements.txt
Pillow
numpy
scikit-image
scipy
opencv-python
google-generativeai
python-dotenv
cairosvg